+++ shell_test.go
@@ -1,7 +1,6 @@
 package cmdline
 
 import (
-	"strings"
 	"testing"
 
 	"github.com/gregoryv/asserter"
@@ -15,41 +14,19 @@ func TestNewShell(t *testing.T) {
 			t.Error("got exit code", v)
 		}
 	}
-	testShell(t, sh)
-}
-
-func TestNewShellT(t *testing.T) {
-	sh := NewShellT()
-	defer sh.Cleanup()
-
-	testShell(t, sh)
-
-	// dump contains exit code
-	got := sh.Dump()
-	if !strings.Contains(got, "exit ") {
-		t.Error(got)
-	}
-
-	// dump can contain content of stderr
-	sh.Err.WriteString("hello")
-	got = sh.Dump()
-	if !strings.Contains(got, "hello") {
-		t.Error(got)
-	}
-}
-
-func testShell(t *testing.T, cmd Shell) {
-	t.Helper()
 	assert := asserter.Wrap(t).Assert
 
-	assert(cmd.Getenv("PWD") != "").Error(`cmd.Getenv("PWD") failed `)
-	assert(len(cmd.Args()) != 0).Error("empty cmd.Args")
+	assert(sh.Getenv("PWD") != "").Error(`sh.Getenv("PWD") failed `)
+	assert(len(sh.Args()) != 0).Error("empty sh.Args")
 
-	wd, _ := cmd.Getwd()
-	assert(wd != "").Error("empty cmd.Getwd()")
-	assert(cmd.Stdin() != nil).Error("nil cmd.Stdin")
-	assert(cmd.Stdout() != nil).Error("nil cmd.Stdout")
-	assert(cmd.Stderr() != nil).Error("nil cmd.Stderr")
-	cmd.Fatal()
-	cmd.Exit(1)
+	wd, err := sh.Getwd()
+	if err != nil {
+		t.Fatal(err)
+	}
+	assert(wd != "").Error("empty sh.Getwd()")
+	assert(sh.Stdin() != nil).Error("nil sh.Stdin")
+	assert(sh.Stdout() != nil).Error("nil sh.Stdout")
+	assert(sh.Stderr() != nil).Error("nil sh.Stderr")
+	sh.Fatal()
+	sh.Exit(1)
 }