+++ changelog.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
-- Add type Shell, ShellOS and ShellT
+- Add types Shell, ShellOS and ShellT
 
 ## [0.7.0] 2021-10-21